+++ core/src/api/v2/handlers.rs
@@ -14,7 +14,10 @@ use crate::{
 			Subscription, SubscriptionId, Transaction, Version, WsClients,
 		},
 	},
-	data::{AppDataKey, BlockHeaderKey, Database, RpcNodeKey, VerifiedCellCountKey},
+	data::{
+		AppDataKey, BlockHeaderKey, BlockHeaderReceivedAtKey, Database, RpcNodeKey,
+		VerifiedCellCountKey,
+	},
 	utils::calculate_confidence,
 };
 
@@ -109,6 +112,10 @@ pub async fn block_header(
 		.and_then(|extension| block_status(sync_start_block, db.clone(), block_number, extension))
 		.ok_or(Error::not_found())?;
 
+	let received_at = db
+		.get(BlockHeaderReceivedAtKey(block_number))
+		.ok_or_else(Error::not_found)?;
+
 	if matches!(
 		block_status,
 		BlockStatus::Unavailable | BlockStatus::Pending | BlockStatus::VerifyingHeader
@@ -118,7 +125,7 @@ pub async fn block_header(
 
 	db.get(BlockHeaderKey(block_number))
 		.ok_or_else(|| eyre!("Header not found"))
-		.and_then(|header| header.try_into())
+		.and_then(|header| (header, received_at).try_into())
 		.map_err(Error::internal_server_error)
 }