+++ core/src/network/rpc/subscriptions.rs
@@ -6,18 +6,18 @@ use avail_rust::{
 use codec::Encode;
 use color_eyre::{eyre::eyre, Result};
 #[cfg(not(target_arch = "wasm32"))]
-use std::time::Instant;
+use std::time::{Instant, SystemTime, UNIX_EPOCH};
 use tokio::sync::broadcast::Sender;
 use tokio_stream::StreamExt;
 use tracing::{debug, info, trace};
 #[cfg(target_arch = "wasm32")]
-use web_time::Instant;
+use web_time::{Instant, SystemTime, UNIX_EPOCH};
 
 use super::{Client, OutputEvent, Subscription};
 use crate::{
 	data::{
-		Database, FinalitySyncCheckpoint, FinalitySyncCheckpointKey, IsFinalitySyncedKey,
-		LatestHeaderKey, VerifiedHeaderKey,
+		BlockHeaderReceivedAtKey, Database, FinalitySyncCheckpoint, FinalitySyncCheckpointKey,
+		IsFinalitySyncedKey, LatestHeaderKey, VerifiedHeaderKey,
 	},
 	finality::{check_finality, ValidatorSet},
 	types::{BlockRange, GrandpaJustification},
@@ -98,8 +98,17 @@ impl<T: Database + Clone> SubscriptionLoop<T> {
 		match subscription {
 			Subscription::Header(header) => {
 				let received_at = Instant::now();
+				let received_at_timestamp = SystemTime::now()
+					.duration_since(UNIX_EPOCH)
+					.expect("Time went backwards")
+					.as_secs();
 				self.db.put(LatestHeaderKey, header.clone().number);
 				info!("Header no.: {}", header.number);
+
+				self.db.put(
+					BlockHeaderReceivedAtKey(header.number),
+					received_at_timestamp,
+				);
 
 				// if new validator set becomes active, replace the current one
 				if self.block_data.next_valset.is_some() {
@@ -163,6 +172,11 @@ impl<T: Database + Clone> SubscriptionLoop<T> {
 				let (header, received_at, valset) =
 					self.block_data.unverified_headers.swap_remove(pos);
 
+				let received_at_timestamp = self
+					.db
+					.get(BlockHeaderReceivedAtKey(header.number))
+					.expect("Block header timestamp is in the database");
+
 				let is_final = check_finality(&valset, &justification);
 
 				is_final.expect("Finality check failed");
@@ -216,6 +230,7 @@ impl<T: Database + Clone> SubscriptionLoop<T> {
 							.send(OutputEvent::HeaderUpdate {
 								header,
 								received_at,
+								received_at_timestamp,
 							})
 							.unwrap();
 					}
@@ -240,6 +255,7 @@ impl<T: Database + Clone> SubscriptionLoop<T> {
 					.send(OutputEvent::HeaderUpdate {
 						header,
 						received_at,
+						received_at_timestamp,
 					})
 					.unwrap();
 			} else {